# Support arbitrary bytes in credentials