--- a/src/Tinfoil/Data/KDF.hs
+++ b/src/Tinfoil/Data/KDF.hs
@@ -30,7 +30,7 @@ instance NFData CredentialHash
 
 newtype Credential =
   Credential {
-    unCredential :: Text
+    unCredential :: ByteString
   } deriving (Eq, Show, Generic)
 
 instance NFData Credential