--- a/test/Test/IO/Tinfoil/Random.hs
+++ b/test/Test/IO/Tinfoil/Random.hs
@@ -5,12 +5,14 @@
 
 module Test.IO.Tinfoil.Random where
 
+import qualified Data.ByteString as BS
 import           Data.Char (ord)
 import           Data.List (intersect)
 import           Data.List.NonEmpty (NonEmpty)
 import qualified Data.List.NonEmpty as NE
 import           Data.Maybe (fromJust)
-import qualified Data.Text                     as T
+import qualified Data.Text as T
+import qualified Data.Text.Encoding as T
 
 import           Disorder.Core.IO
 
@@ -27,12 +29,12 @@ import           Test.Tinfoil.Arbitrary
 prop_randomCredential_length :: Property
 prop_randomCredential_length = forAll ((,) <$> credentialLength <*> excludedChars) $ \(l, ex) -> testIO $ do
   c <- randomCredential ex l
-  pure $ (T.length . unCredential $ fromJust c) === l
+  pure $ (BS.length . unCredential $ fromJust c) === l
 
 prop_randomCredential_exclusion :: Property
 prop_randomCredential_exclusion = forAll ((,) <$> credentialLength <*> excludedChars) $ \(l, ex) -> testIO $ do
   c <- randomCredential ex l
-  pure $ (intersect (T.unpack . unCredential $ fromJust c) ex) === []
+  pure $ (intersect (T.unpack . T.decodeUtf8 . unCredential $ fromJust c) ex) === []
 
 prop_randomCredential_empty :: Property
 prop_randomCredential_empty = forAll credentialLength $ \l -> testIO $ do
@@ -42,7 +44,7 @@ prop_randomCredential_empty = forAll credentialLength $ \l -> testIO $ do
 prop_randomCredential_printable :: Property
 prop_randomCredential_printable = forAll credentialLength $ \l -> testIO $ do
   cr <- randomCredential [] l
-  pure $ all (\c -> ord c < 128 && ord c >= 32) (T.unpack . unCredential $ fromJust cr)
+  pure $ all (\c -> ord c < 128 && ord c >= 32) (T.unpack . T.decodeUtf8 . unCredential $ fromJust cr)
 
 prop_drawOnce_closure :: Property
 prop_drawOnce_closure = forAll arbitrary $ \(xs :: NonEmpty Int) -> testIO $ do