Required for `eminence`.

Also bump IO test iterations to fit with the guidelines.

@markhibberd 

Fixes #17.